requirements.txt
flask
tensorflow
tflite-runtime
pillow
numpy
pandas